09 add in tests for xOf scenarios
